refactor(db): extract connection schema setup
